requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
colour-science==0.3.16
lightning-bolts==0.4.0
opencv-python==4.5.5.62
openpyxl==3.0.7
pandas==1.1.5
pytorch-lightning==1.4.2
setuptools==59.5.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchmetrics==0.5.1
torchvision==0.10.0+cu111